refactor: add sidebar component
